refactor: 🧩 Change apiRequestManager to requestUtils
