+++ src/collections/requests/anilist.js
