+++ src/collections/requests/jikan.js
