+++ src/utils/utils.js
@@ -5,4 +5,4 @@ export * as fetcherUtils from "./Fetcher/fetcherUtils.js";
 export * as scheduleUtils from "./scheduled.js";
 export * as dateUtils from "./dateUtils.js";
 
-export * as apiRequestManager from "./rateLimits/requests.js";
+export * as requestUtils from "./requestUtils.js";
